[#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동
- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함 [#34] refactor: getColor 함수를 Calendar 컴포넌트 안으로 이동

- getColor 함수가 캘린더에서만 쓰이기 때문에 그 안으로 이동함
