+++ FE/src/components/CalendarModal/Calendar/getColor.js
